refactor: Split nodes/identity.ts from nodes/Node.ts
