+++ src/nodes/YAMLMap.ts
@@ -4,7 +4,8 @@ import type { StringifyContext } from '../stringify/stringify.js'
 import { stringifyCollection } from '../stringify/stringifyCollection.js'
 import { addPairToJSMap } from './addPairToJSMap.js'
 import { Collection } from './Collection.js'
-import { isPair, isScalar, MAP, ParsedNode, Range } from './Node.js'
+import { isPair, isScalar, MAP } from './identity.js'
+import type { ParsedNode, Range } from './Node.js'
 import { Pair } from './Pair.js'
 import { isScalarValue, Scalar } from './Scalar.js'
 import type { ToJSContext } from './toJS.js'
