+++ src/nodes/YAMLSeq.ts
@@ -3,7 +3,8 @@ import type { Schema } from '../schema/Schema.js'
 import type { StringifyContext } from '../stringify/stringify.js'
 import { stringifyCollection } from '../stringify/stringifyCollection.js'
 import { Collection } from './Collection.js'
-import { isScalar, ParsedNode, Range, SEQ } from './Node.js'
+import { isScalar, SEQ } from './identity.js'
+import type { ParsedNode, Range } from './Node.js'
 import type { Pair } from './Pair.js'
 import { isScalarValue, Scalar } from './Scalar.js'
 import { toJS, ToJSContext } from './toJS.js'
